removed unneeded functions
